WIP: moves Project to /editor; issues temporary import fixes
