+++ src/views/Create/Create.jsx
@@ -1,6 +1,6 @@
 import styles from './Create.css';
-import { ProjectProvider } from '../../context/ProjectContext';
-import Project from '../../components/Project/Project';
+import { ProjectProvider } from '@context/ProjectContext';
+import Project from '@components/editor/Project/Project';
 
 export default function Create() {
   return (
